backend/src/api/requirements_render.txt
# backend\src\api\requirements_render.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
